queue switching bug fixed, added folder structure

+++ Sim7/src/nQueues/CustomerArrivalEvent.java
@@ -23,7 +23,7 @@ public class CustomerArrivalEvent extends Event<Customer>{
 		
 		if(random) {
 			// random machine
-			machine = model.getQueue();
+			machine = (int) model.getQueue();
 		}else {
 			// find the best queue and hence the machine
 			for(int i=1; i<numberOfMachines; i++){
